* Replaced some custom choose player effects.
